Reinstall system from packages during sysa->sysc and sysb->sysc transitions

+++ sysa/linux-4.9.10/linux-4.9.10.sh
@@ -8,7 +8,7 @@ src_unpack() {
     mv "${DISTFILES}/deblob-4.9" "${pkg}/"
     # Remove all previous source tarballs
     mv "${DISTFILES}/${pkg}.tar.gz" .
-    rm -r "${DISTFILES}/"*
+    rm -r "${DISTFILES:?}/"*
     mv "${pkg}.tar.gz" "${DISTFILES}"
     default || true # Predictable link errors - not a problem
     rm -r "${DISTFILES}" # Clear storage space
